+++ flutter_mintegral-1.5.1/flutter_mintegral/lib/flutter_mintegral.dart
@@ -1,8 +1,0 @@
-
-import 'flutter_mintegral_platform_interface.dart';
-
-class FlutterMintegral {
-  Future<String?> getPlatformVersion() {
-    return FlutterMintegralPlatform.instance.getPlatformVersion();
-  }
-}